requirements.txt
altgraph==0.17.5
packaging==25.0
pefile==2024.8.26
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
python-vlc==3.0.21203
pywin32-ctypes==0.2.3
setuptools==80.9.0
